Implement SendLocation method

+++ type.go
@@ -261,8 +261,8 @@ type Voice struct {
 
 // Location represents a point on the map.
 type Location struct {
-	Lat  float32 `json:"latitude"`
-	Long float32 `json:"longitude"`
+	Lat  float64 `json:"latitude"`
+	Long float64 `json:"longitude"`
 }
 
 // Contact represents a phone contact.
